requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.3
bleach==6.0.0
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-log==0.4.0
colorama==0.4.6
dataclasses==0.6
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
docutils==0.20.1
dotty-dict==1.3.1
exceptiongroup==1.2.0
Flask==2.2.5
gitdb==4.0.11
GitPython==3.1.42
gotrue==0.5.4
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.3
joblib==1.3.2
keyring==24.1.1
langdetect==1.0.9
MarkupSafe==2.1.5
more-itertools==9.1.0
numpy
openai==1.29.0
packaging==24.0
pipreqs==0.4.13
pkginfo==1.10.0
playwright==1.35.0
postgrest==0.10.4
postgrest-py==0.10.4
pydantic==1.10.14
pydantic_core==2.14.6
pyee==9.0.4
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-gitlab==3.15.0
python-semantic-release==7.32.1
readme-renderer==37.3
realtime==0.0.5
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.4.1
stem==1.8.2
storage3==0.3.6
StrEnum==0.4.15
supabase==0.7.1
supafunc==0.2.2
threadpoolctl==3.1.0
tomlkit==0.12.4
tqdm==4.66.2
twine==3.8.0
typing_extensions==4.7.1
urllib3==2.0.7
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3
yarg==0.1.9
zipp==3.15.0
pycryptodome
PyPDF2
